% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{cohort_to_calendar}
\alias{cohort_to_calendar}
\title{Map lifetimes to calendar days}
\usage{
cohort_to_calendar(lifetimes, start_day = 1)
}
\arguments{
\item{lifetimes}{a numeric vector of lifetime durations}

\item{start_day}{a length one numeric vector allowing new cohorts to be added in a ragged fashion (as they come)}
}
\value{
tibble::tibble
}
\description{
Map lifetimes to calendar days
}
