% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{subscribers_by_day}
\alias{subscribers_by_day}
\title{Simulation of subscribers}
\usage{
subscribers_by_day(dist = c("weibull", "lnorm"), ...)
}
\arguments{
\item{dist}{the distribution characterizing churn (weibull and lnorm are implemented)}

\item{...}{parameters that complete the characterization of churn, like shape = 0.8}
}
\value{
tibble::tibble
}
\description{
Simulation of subscribers
}
\examples{
subscribers_by_day("weibull", 0.8, 300)
}
