% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{subscribers_by_day}
\alias{subscribers_by_day}
\title{Simulation of subscribers}
\usage{
subscribers_by_day(name, rng)
}
\arguments{
\item{name}{a length one character vector describing the name of the simulation}

\item{rng}{a random number generator reflecting subscription lifetimes}
}
\value{
tibble::tibble
}
\description{
Simulation of subscribers
}
\examples{
subscribers_by_day("My experiment", function(x) { rweibull(x, 0.8, 500) })
}
