% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\name{subscribers_by_day}
\alias{subscribers_by_day}
\title{Simulation of subscribers}
\usage{
subscribers_by_day(cohorts, dist = c("weibull", "lnorm"), ...)
}
\arguments{
\item{cohorts}{an integer vector of the count of subscriptions sold per day, assumed continuous from an origin day 1.}

\item{dist}{the distribution characterizing churn (weibull and lnorm are implemented)}

\item{...}{parameters that complete the characterization of churn, like shape = 0.8}
}
\value{
tibble::tibble
}
\description{
Simulation of subscribers
}
\examples{
# rep(1e2, 10) is 100 subscriptions simulated for for ten days.
subscribers_by_day(cohorts = rep(1e2, 100), dist = "weibull", 0.8, 300)
}
